+++ src/Command/LoadSingleFixtureCommand.php
@@ -8,6 +8,7 @@ use Basecom\FixturePlugin\FixtureLoader;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
 
@@ -27,6 +28,7 @@ class LoadSingleFixtureCommand extends Command
     {
         $this
             ->setHelp('This command allows you to run only one specific fixture')
+            ->addOption('with-dependencies', 'w', InputOption::VALUE_NONE, 'Run fixture with dependencies')
             ->addArgument('fixtureName', InputArgument::REQUIRED, 'Name of Fixture to load');
     }
 
@@ -44,7 +46,14 @@ class LoadSingleFixtureCommand extends Command
             return Command::FAILURE;
         }
 
-        $this->loader->runSingle($io, $groupNameInput);
+        $withDependencies = $input->getOption('with-dependencies');
+        if (!\is_bool($withDependencies)) {
+            $io->error('Please make sure that your argument is of type boolean');
+
+            return Command::FAILURE;
+        }
+
+        $this->loader->runSingle($io, $groupNameInput, $withDependencies);
 
         $io->success('Done!');